requirements.txt
appdirs==1.4.4
asgiref==3.5.0
attrs==21.4.0
black==21.7b0
certifi==2021.10.8
charset-normalizer==2.0.10
click==8.0.0
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
coverage==6.3
dj-database-url==0.5.0
Django==3.2.7
django-cors-headers==3.8.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
drf-compose==0.1.1
drf-yasg==1.20.0
emoji==1.4.2
flake8==4.0.1
idna==3.3
inflection==0.5.1
iniconfig==1.1.1
isort==5.10.1
itypes==1.2.0
Jinja2==3.0.3
MarkupSafe==2.0.1
mccabe==0.6.1
mypy-extensions==0.4.3
packaging==21.3
pathspec==0.9.0
pluggy==1.0.0
psycopg2-binary==2.9.3
py==1.11.0
py-sync-dotenv==0.1.1
pycodestyle==2.8.0
pyflakes==2.4.0
PyJWT==2.3.0
pyparsing==3.0.7
pytest==6.2.5
pytest-django==4.5.2
pytest-mock==3.6.1
python-dotenv==0.19.2
pytz==2021.3
PyYAML==5.4.1
regex==2022.1.18
requests==2.27.1
ruamel.yaml==0.17.20
ruamel.yaml.clib==0.2.6
sqlparse==0.4.2
toml==0.10.2
tomli==1.2.3
uritemplate==4.1.1
urllib3==1.26.8
watchdog==2.1.6
